Move single-use MCP helpers
Move manager-only startup update, origin, and startup error display helpers from mcp_connection.rs into manager.rs.

Move client-only bearer token resolution and server-name validation helpers into client.rs.
